% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leak.R
\name{accessLeakEnv}
\alias{accessLeakEnv}
\title{Accessor function to return the hidden leakEnv (mainly for testing)}
\usage{
accessLeakEnv()
}
\description{
Accessor function to return the hidden leakEnv (mainly for testing)
}
