% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{add.error.ac}
\alias{add.error.ac}
\title{the result of this function should be multiplied by a variable to add
log normal error to a value}
\usage{
add.error.ac(ac.err, evar)
}
\arguments{
\item{ac.err}{result of a call to get.error.ac}

\item{evar}{is the variance in your error term}
}
\description{
the result of this function should be multiplied by a variable to add
log normal error to a value
}
\examples{
new.ac<-get.error.ac(ac,old.ac,management.error)
old.ac<-new.ac   #keep track of the previous value for next time step
SSBapparent<-SSB*add.error.ac(new.ac,management.error)
}
