% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{add.error}
\alias{add.error}
\title{Add error to a simulated process.}
\usage{
add.error(evar)
}
\arguments{
\item{evar}{the log space variance in your error term}
}
\description{
The result of this function should be multiplied by your variable to add
log normal error to a value.
}
\examples{
FullF=c(0.2,0.22,0.25);F.error=0.2
Fapparent<-FullF*add.error(F.error)
print(data.frame("Full.F"=FullF,"FwithError"=Fapparent))

}
