% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Population.R
\name{add_Population}
\alias{add_Population}
\title{Add a population to a rule book}
\usage{
add_Population(book, populations = 1)
}
\arguments{
\item{book}{\code{Character} value, name of the rule book to be modified.}

\item{populations}{\code{Numeric} value, number of additional populations
to create.}
}
\value{
A list object with all rules for a model run.
}
\description{
The function adds a further population element to all rules or a rule book.
}
\examples{

## create simple true age-depth-relationship
book_1 <- get_RuleBook(book = "empty")

book_2 <- add_Population(book = book_1, 
                         populations = 1)
                          
}
\author{
Michael Dietze
}
