% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Rule.R
\name{add_Rule}
\alias{add_Rule}
\title{Add a rule to a rule book}
\usage{
add_Rule(book, name, group, type, populations = 1)
}
\arguments{
\item{book}{\code{Character} value, name of the rule book to be modified.}

\item{name}{\code{Character} value, name of the rule rule to be added.}

\item{group}{\code{Character} value, group to which the rule belongs. One
out of \code{"general"} (covering the sediment section properties) and
\code{"specific"} (relevant for a single grain).}

\item{type}{\code{Character} value, generic type of the rule. One out of
\code{"exact"} (defined by exact value, changing with depth),
\code{"normal"} (normal distribution, defined by mean and standard
deviation, changing with depth), \code{"uniform"} (defined by min and
max value, changing with depth) and \code{"gamma"} (gamma distribution,
defined by shape and scale parameter and constant offset, all changing
with depth)}

\item{populations}{\code{Numeric} value, number of populations to create.
The number of populations to add should match the existing number of
populations.}
}
\value{
A list object with all rules for a model run.
}
\description{
The function adds a new rule to an existing rule book. The specified rule
will be appended to the rule book.
}
\examples{

## create simple true age-depth-relationship
book_1 <- get_RuleBook()

book_2 <- add_Rule(book = book_1, 
                          name = "extrarule", 
                          group = "general", 
                          type = "normal", 
                          populations = 1)
                
}
\author{
Michael Dietze
}
