% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.r, R/add2.r
\name{add_x}
\alias{add_x}
\title{Add once}
\usage{
add_x(x)

add_x(x)
}
\arguments{
\item{x}{Double or Integer}

\item{x}{Double or Integer}
}
\value{
x+x

x+2
}
\description{
Adds the input to itself

Adds 2 to the input
}
\examples{
add_x(2)

add_x(2)

}
