\name{altmetmaster}
\alias{altmetmaster}
\title{Plot altmetrics}
\usage{
  altmetmaster(altmet = "10.1038/480426a",
    gitid = "cboettig", gittype = "user",
    bitlyurl = "http://bit.ly/xaTrO7")
}
\arguments{
  \item{diggname}{Digg username}

  \item{gitid}{GitHub username}

  \item{gittype}{GitHub type, 'user' or 'org'}
}
\description{
  This function plots metrics on persons or articles from
  many sources.
}
\examples{
altmetmaster()
}
\author{
  Scott Chamberlain
}

