% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{asm.ci}
\alias{asm.ci}
\title{generate approximate confidence intervals based on lognormal variable x}
\usage{
asm.ci(x, cv.x, bounds = 95)
}
\arguments{
\item{bounds}{default=95 the width of your confidence intervals}

\item{ts}{is your data}

\item{cv}{its cv}
}
\description{
generate approximate confidence intervals based on lognormal variable x
}
