% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_fdr.R
\name{bh_fdr}
\alias{bh_fdr}
\title{False Discovery Threshold}
\usage{
bh_fdr(pvals, FDR = 0.05)
}
\arguments{
\item{pvals}{Vector of p-values}

\item{FDR}{False-discovert rate}
}
\description{
Determines the critical p-value for statistical significance using the Benjamini-Hochberg false-discovery rate procedure.

NOTE: Significance is defined as P <= threshold, not <.
}
