% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{bind_disabled}
\alias{bind_disabled}
\title{Bind Disabled State}
\usage{
bind_disabled(tag, expr)
}
\arguments{
\item{tag}{An htmltools tag.}

\item{expr}{A JavaScript expression string referencing the model
(e.g. \code{"model.count"}).}
}
\value{
The modified tag.
}
\description{
Dynamically disables a form element based on a JavaScript
expression.
}
\examples{
shiny::tags$button("Submit") |>
  bind_disabled("model.loading")

}
