% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{bind_style}
\alias{bind_style}
\title{Bind Dynamic Style}
\usage{
bind_style(tag, expr)
}
\arguments{
\item{tag}{An htmltools tag.}

\item{expr}{A JavaScript expression string referencing the model
(e.g. \code{"model.count"}).}
}
\value{
The modified tag.
}
\description{
Dynamically sets inline styles from a JavaScript expression.
}
\examples{
shiny::tags$div(class = "progress-bar") |>
  bind_style("'width: ' + model.progress + '\%'")

}
