% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{blank.plot}
\alias{blank.plot}
\title{plot nothing but string in the middle of the plotting region - a useful
placeholder for automated plotting loops}
\usage{
blank.plot(string)
}
\arguments{
\item{string}{quoted string you want displyed at the center of the blank plot}
}
\description{
plot nothing but string in the middle of the plotting region - a useful
placeholder for automated plotting loops
}
