% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAggregateBiomass.R
\name{calcAggregateBiomass}
\alias{calcAggregateBiomass}
\title{Calculate Aggregate Biomass}
\usage{
calcAggregateBiomass(NAA, WAA, Mval = 0, Fval = 0, indextime = 0)
}
\arguments{
\item{NAA}{vector of numbers at age (e.g., PopN, Catch, Index)}

\item{WAA}{vector of weights at age}

\item{Mval}{vector of natural mortality at age or single value applied to all ages, default = 0}

\item{Fval}{vector of fishing mortality at age or single value applied to all ages, default = 0}

\item{indextime}{proportion of year that occurs before index generated, default = 0}
}
\description{
Multiplies vector of numbers at age by weights at age and adjusts for time of year.
}
