% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSurvival.R
\name{calcSurvival}
\alias{calcSurvival}
\title{Cohort Survival}
\usage{
calcSurvival(PopN, Mval, Fval, plusgroupflag = TRUE)
}
\arguments{
\item{PopN}{population abundance, can be number or vector}

\item{Mval}{natural mortality rate, can be number or same dimension as PopN}

\item{Fval}{fishing mortality rate, can be number or same dimension as PopN}

\item{plusgroupflag}{means last element in vector or matrix is a plus group, default = TRUE}
}
\description{
Calculate survival of a cohort with optional plus group calculation given PopN, M, and F.
}
