% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captureleakedobjects.R
\name{captureLeakedObjects}
\alias{captureLeakedObjects}
\title{Return leaked objects from sandbox environment to the parent}
\usage{
captureLeakedObjects(x, env = parent.frame())
}
\arguments{
\item{x}{object to return from sandbox environment}

\item{env}{environment to assign the objects stored in x to}
}
\description{
Return leaked objects from sandbox environment to the parent
}
