% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{castSand}
\alias{castSand}
\title{Evaluate code from original R Session}
\usage{
castSand(mold)
}
\arguments{
\item{mold}{code supplied to sandbox R session to evaluate, contained in a function call}
}
\description{
Evaluate code from original R Session
}
