% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_Utils.R
\name{castSand_serial}
\alias{castSand_serial}
\title{Evaluate code from a serialized function}
\usage{
castSand_serial(mold)
}
\arguments{
\item{mold}{string version of the serialized function to evaluate}
}
\description{
Evaluate code from a serialized function
}
