\name{ChangeLog}
\title{Change Log for scABC2}
\description{

  \strong{v.0.0.3}:
  
  \itemize{
    \item added fx.return_axis_length to handle X and Y axis coordinates in both
    1D vectors as well a matrices and data.frames, which would return a different
    value when used with length(df)
    \item fixed bug in plt.DimRedPlot: "y_axs" had been used as the default for
    both X and Y
    }
  
  
}
