% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{check.package}
\alias{check.package}
\title{useful for getting packages in functions}
\usage{
check.package(package1)
}
\arguments{
\item{package1}{duh}
}
\description{
useful for getting packages in functions
}
