% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check user input in a given context}
\usage{
check(expr, id)
}
\arguments{
\item{expr}{an expression}

\item{id}{task id}
}
\description{
Check user input in a given context
}
\examples{
\dontrun{
check(x + y, 1)
}
}
