% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkOMdims.R
\name{checkOMdims}
\alias{checkOMdims}
\title{Check Operating Model Dimensions}
\usage{
checkOMdims(years, nages, nindices, nyear1list, Rlist, Mlist, Flist, Wlist,
  indexlist, catcherrorlist, indexerrorlist, processerrorlist = FALSE)
}
\arguments{
\item{years}{vector of years in operating model (OM)}

\item{nages}{number of ages in OM (assumed to start at age 1)}

\item{nindices}{number of indices}

\item{nyear1list}{N in first year list}

\item{Rlist}{recruitment list}

\item{Mlist}{natural mortality list}

\item{Flist}{fishing mortality list}

\item{Wlist}{weights at age list}

\item{indexlist}{list of index lists}

\item{catcherrorlist}{list for observation errors in catch}

\item{indexerrorlist}{list of lists for observation errors in the indices}

\item{processerrorlist}{list for process error in survival equation, default=FALSE}
}
\description{
Check that all dimensions of the operating model are consistent (e.g., number of years, ages, indices). Use all(unlist(checklist) == TRUE) to see if all checks passed.
}
