% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{closeSandbox}
\alias{closeSandbox}
\title{Quit from sandboxed R session}
\usage{
closeSandbox(sandboxCon)
}
\arguments{
\item{socketCon}{socket connection to original R session}
}
\description{
Quit from sandboxed R session
}
