% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_fibonacci.R
\name{collapse_fibonacci}
\alias{collapse_fibonacci}
\title{Return a length 1 vector encompassing the first n elements of the Fibonacci sequence}
\usage{
collapse_fibonacci(n, sep = "")
}
\arguments{
\item{n}{Numeric, integer, indicating how many elements to use}

\item{sep}{Character element to use as a separator}
}
\value{
Character, vector of length 1, encompassing the first n elements of the Fibonacci sequence, seprated by the separator
}
\description{
Return a length 1 vector encompassing the first n elements of the Fibonacci sequence
}
\examples{
collapse_fibonacci(5)
collapse_fibonacci(3, sep = "_")


}
