% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{Convert between phi units and micrometers}
\usage{
convert_units(phi, mu)
}
\arguments{
\item{phi}{\link{numeric} vector, grain-size class values
in phi to be converted}

\item{mu}{\link{numeric} vector, grain-size class values
in micrometres to be converted}
}
\value{
\link{numeric} vector, converted grain-size class values
}
\description{
The function converts values from the phi-scale to
the micrometer-scale and vice versa.
}
\examples{
## load example data set
## generate phi-values
phi <- -2:5

## convert and show phi to mu
mu  <- convert_units(phi = phi)
mu

## convert and show mu to phi
convert_units(mu = mu)

}
\author{
Michael Dietze, GFZ Potsdam (Germany)
}
