% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deselect.R
\name{deselect}
\alias{deselect}
\title{Deselect variables by name}
\usage{
deselect(.data, ...)
}
\arguments{
\item{.data}{Data frame.}

\item{...}{One or more unquoted expressions separated by commas. You can treat variable names like they are positions, so you can use expressions like x:y to select ranges of variables.

Positive values drop variables; negative values keep variables.}
}
\value{
An object of the same class as .data.
}
\description{
Choose variables from a tbl. deselect() drops only the variables you mention. Supports \code{tidyselect::select_helpers}
}
\examples{

iris \%>\%
  deselect(Species)

}
