% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicom2nifti.R
\name{dicom2nifti}
\alias{dicom2nifti}
\title{Convert a DICOM series to Nifti}
\usage{
dicom2nifti(dicom_dir, nifti_dir, ...)
}
\arguments{
\item{dicom_dir}{Directory containing DICOM series}

\item{nifti_dir}{Directory to save nifti file}

\item{...}{Arguments to pass on to \code{\link{oro.dicom}}}
}
\description{
This function will accept a directory containing a single DICOM
    series as the input, and will return a single nifti file (.nii)
}
\author{
Matthew T. Warkentin (warkentin@lunenfeld.ca)
}
