% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effect_notify}
\alias{effect_notify}
\title{Create a Notification Effect}
\usage{
effect_notify(text, type = "default", duration = 5)
}
\arguments{
\item{text}{Notification text.}

\item{type}{One of \code{"default"}, \code{"message"}, \code{"warning"}, \code{"error"}.}

\item{duration}{Seconds to display. Use \code{NULL} for persistent.}
}
\value{
An \code{mvu_effect} object.
}
\description{
Describes a \code{\link[shiny:showNotification]{shiny::showNotification()}} call to be executed by the
runtime after the model update.
}
\examples{
effect_notify("File saved", type = "message", duration = 3)

}
