% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateSandbox.R
\name{evaluateSandbox}
\alias{evaluateSandbox}
\title{Make Sandbox Run}
\usage{
evaluateSandbox(node, ID, host = "local")
}
\arguments{
\item{node}{the host server to execute code}

\item{ID}{the port ID to open socket on}

\item{host}{location of host server, local or external}
}
\description{
wrapper function to assist with running in sandbox
}
