% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{fac_to_string}
\alias{fac_to_string}
\title{get rid of factors in a data frame}
\usage{
fac_to_string(dataframe)
}
\arguments{
\item{dataframe}{self explanatory I hope}
}
\description{
get rid of factors in a data frame
}
