% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor-ref-level.R
\name{fct_declare}
\alias{fct_declare}
\title{Declare and Recover Factor Reference Level}
\usage{
fct_declare(data, ..., .sep = "")
}
\arguments{
\item{data}{Data frame.}

\item{...}{Factor variables to declare.}

\item{.null}{Null effect size.}
}
\value{
The input data is returned with attributes attached.
}
\description{
Text goes here...
}
\details{
Text here...
}
