% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fibonacci}
\alias{fibonacci}
\title{Vector with Fibonacci sequence}
\format{
Numeric vector with 9 elements
}
\usage{
fibonacci
}
\description{
A numeric vector with the first few elements of the Fibonacci sequence
}
\keyword{datasets}
