% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendingFibonacci.R
\name{fibonacci_Extend}
\alias{fibonacci_Extend}
\title{Extend the Fibonacci vector}
\usage{
fibonacci_Extend(n)
}
\arguments{
\item{n, }{length of the extension}
}
\value{
the extended vector
}
\description{
Extend the Fibonacci vector
}
\examples{
fibonacci_Extend(10)
}
