% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci_log_n.R
\name{fibonacci_log_n}
\alias{fibonacci_log_n}
\title{Return logarithm in base n}
\usage{
fibonacci_log_n(n)
}
\arguments{
\item{n}{Numeric, integer, indicating what base to use for the log function}
}
\value{
Numeric, vector which is the n-based log of fibonacci
}
\description{
Return logarithm in base n
}
\examples{
fibonacci_log_n(10)
fibonacci_log_n(exp(1))
fibonacci_log_n(2)

}
