% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fibonacci_mat}
\alias{fibonacci_mat}
\title{Matrix of values to use as test data}
\format{
Matrix with 100 rows and 9 columns
}
\usage{
fibonacci_mat
}
\description{
A matrix, where column n contains a random sample from a normal
distribution where the mean is the nth element of the fibonacci vector
}
\keyword{datasets}
