% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci_multiply_n.R
\name{fibonacci_multiply_n}
\alias{fibonacci_multiply_n}
\title{Multiply the first n values in fibonacci}
\usage{
fibonacci_multiply_n(n)
}
\arguments{
\item{n}{Numeric, index of the last value in fibonacci to recursively multiply}
}
\value{
Numeric, cumulative product of the first n values of fibonacci
}
\description{
Multiply the first n values in fibonacci
}
\examples{
fibonacci_multiply_n(5)
}
