% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{fibonaci_divide_n}
\alias{fibonaci_divide_n}
\title{divide fibonacci sequence by constant n}
\usage{
fibonaci_divide_n(n)
}
\arguments{
\item{n}{Numberic, number to divide each element in sequence by}
}
\description{
divide fibonacci sequence by constant n
}
\examples{
fibbonacci_divide_n(2)
}
