% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{fill}
\alias{fill}
\title{a function to make a vector out of a scaler and a matrix out of a vector.
A little ticky when you e.g. need to get 1000 years out
of a time varying vector - this will hold the last year's values constant for the remaining years}
\usage{
fill(x, n, m)
}
\arguments{
\item{x}{the scaler or vector you want to expand}

\item{n}{number of rows}

\item{m}{number of cols}
}
\description{
a function to make a vector out of a scaler and a matrix out of a vector.
A little ticky when you e.g. need to get 1000 years out
of a time varying vector - this will hold the last year's values constant for the remaining years
}
