% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{filled.legend}
\alias{filled.legend}
\title{modification of filled.contour by Carey McGilliard and Bridget Ferris
designed to just plot the legend}
\usage{
filled.legend(x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1, length.out
  = ncol(z)), z, xlim = range(x, finite = TRUE), ylim = range(y, finite =
  TRUE), zlim = range(z, finite = TRUE), levels = pretty(zlim, nlevels),
  nlevels = 20, color.palette = cm.colors,
  col = color.palette(length(levels) - 1), plot.title, plot.axes, key.title,
  key.axes, asp = NA, xaxs = "i", yaxs = "i", las = 1, axes = TRUE,
  frame.plot = axes, ...)
}
\description{
modification of filled.contour by Carey McGilliard and Bridget Ferris
designed to just plot the legend
}
