% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dimRedPlot_fx.R
\name{fx.add_cellInfo_to_redDim_df}
\alias{fx.add_cellInfo_to_redDim_df}
\title{Add metadata about the cells (columns) to the data.frame for plotting}
\usage{
fx.add_cellInfo_to_redDim_df(object, df_to_plot, what_info, max_levels = NULL)
}
\arguments{
\item{object}{\code{SingleCellExperiment} object}

\item{df_to_plot}{\code{data.frame} for ggplot2-based plotting}

\item{what_info}{vector of names that will be used to extract the
corresponding values from \code{object}}

\item{max_levels}{indicate a number if you want to limit the number of levels
that any of the entries specified by \code{what_info} can have. If the
number is exceeded, the entry will be dropped. Default: NULL}
}
\value{
\code{df_to_plot} with additional columns
}
\description{
Add metadata about the cells (columns) to the data.frame for plotting
}
\examples{
\dontrun{#' 
df_to_plot <- fx.add_cellInfo_to_redDim_df(sce, df_to_plot, names(colData(sce))[1:4])
}

}
