% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dimRedPlot_fx.R
\name{fx.add_factors_to_redDim_df}
\alias{fx.add_factors_to_redDim_df}
\title{Adding color values to a data.frame for plotting reduced dimensions}
\usage{
fx.add_factors_to_redDim_df(object, df_to_plot, color_by = NULL,
  shape_by = NULL, size_by = NULL, circle_by = NULL,
  exprs_values = NULL)
}
\arguments{
\item{object}{\code{SCE} or \code{SCESet} object}

\item{color_by}{entry of either \code{colData(object)},
\code{rowData(object)}, \code{rownames(object)},
which will be used to assign either discrete or continous color schemes.
Alternatively, you can supply a \code{list} with a label stored in \code{list$title}
and values stored in \code{list$result} (\code{result} should be as long as
\code{dim(object)[2]}. Default: \code{NULL}}

\item{shape_by}{default: \code{NULL}}

\item{size_by}{default: \code{NULL}}

\item{circle_by}{Used for encircling points of interest; default: \code{NULL}}

\item{exprs_values}{The name of the expression values which will be used for
the legend of the plot.}

\item{dt_to_plot}{\code{data.frame} fit for ggplot2-based routines}
}
\value{
\code{data.frame} fit for plotting with \code{ggplot2} with
additional columns for color, shape and size
}
\description{
Adding color values to a data.frame for plotting reduced dimensions
}
\details{
Basically a wrapper around \code{\link{fx.choose_vis_values_sce}}
that already makes smart assumptions about reasonable checks for any of
the factors specified via \code{size_by}, \code{color_by}, \code{shape_by}.
This function will add columns for those factors so that they can be
conveniently used for faceting etc. using \code{ggplot2}.
}
\seealso{
\code{\link{fx.choose_vis_values_sce}}
}
