% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dimRedPlot_fx.R
\name{fx.choose_vis_values_sce}
\alias{fx.choose_vis_values_sce}
\title{Add visualization factors for ggplot2}
\usage{
fx.choose_vis_values_sce(x, by, check_coldata = TRUE,
  cell_control_default = FALSE, check_features = FALSE,
  exprs_values = "counts", coerce_factor = FALSE, level_limit = NA)
}
\arguments{
\item{x}{\code{SingleCellExperiment} object}

\item{by}{string indicating which factor will be checked}

\item{check_coldata}{whether \code{colData(object)}
should be checked for values of \code{by}; default: TRUE}

\item{check_features}{whether \code{featureNames(object)} or \code{rownames(object)}
should be checked for values of \code{by}; default: FALSE, which saves time}

\item{exprs_values}{which type of expression values will be used if \code{by}
is a feature}

\item{level_limit}{maximum numbers of levels that are going to be allowed;
usually only applicable if \code{by} defines a shape}
}
\value{
a list with the label of the factor (\code{name}) as well as the actual
values that are going to be used for the plotting
}
\description{
Add visualization factors for ggplot2
}
\details{
This function looks through the visualization data
and returns the values to be visualized. Either \code{by}
itself, or a column of \code{colData}, or a column of \code{rowData},
or the expression values of a feature.
}
\seealso{
\code{\link{fx.add_factors_to_redDim_df}}
}
