% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drp.R
\name{fx.get_axes_val_names}
\alias{fx.get_axes_val_names}
\title{Extract the column names that correspond to the X and Y axis}
\usage{
fx.get_axes_val_names(in.df, X = "x_axs", Y = "y_axs")
}
\arguments{
\item{in.df}{Data.frame whose names() will be used to extract the instance that
matches \code{X} and \code{Y}}

\item{X}{name or integer index of the column that will be used for X axis
values. The default is "x_axs".}

\item{Y}{name or integer index of the column that will be used for Y axis
values. The default is "y_axs".}
}
\value{
vector of names that correspond to the values where X_regex and Y_regex
match.
}
\description{
Very simple function that's basically just an error check for
whether the regex we think should be extracting the names of columns corresponding
to X and Y values work as expected: there should only be two values returned.
}
\examples{
\dontrun{

## using column names
fx.get_axes_val_names(test.df, X = "PC1", Y = "PC2")

# using integer indeces
fx.get_axes_val_names(test.df, X = 1L, Y = 2L)

}
}
