% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drp.R
\name{fx.get_circle_df}
\alias{fx.get_circle_df}
\title{Make a data.frame for encircling points in a reduced dimension plot}
\usage{
fx.get_circle_df(in.df, X = "x_axs", Y = "y_axs", separate_by = NULL,
  keep_quant = 1)
}
\arguments{
\item{in.df}{data.frame that is used for making the original plot.}

\item{X}{name (or integer index) of the column of in.df that contains the 
values to be used for determining the x axis coordinates (default: "x_axs").}

\item{Y}{name (or integer index) of the column that contains the values to be 
used for determining the y axis coordinates (default: "y_axs".,}

\item{separate_by}{factor that will be used to split the values of X and Y
into distinct groups (e.g., by "condition"); if NULL (default), \code{which_quant}}

\item{keep_quant}{a number between 0.5 and 1 to define the max. quantile for
any value of \code{X} or \code{Y} that will be kept for defining the circle
coordinates. The default is to capture all points that belong to one instance
of \code{separate_by}, i.e. \code{keep_quant= 1}. It might be useful to set it
to less than 1 (e.g., 0.99) if you don't want the circle to be blown up by
individual outliers. 0.99 would ensure that only those points between the 99th
percentile and the 0.01st percentile (1 - 0.99) would be kept.}
}
\value{
data.table, most likely a subset of the in.df, depending on the setting
of \code{which_quant}. This data.table can be used with ggalt::geom_encircle.
}
\description{
For the encircling, we will use the values of X and Y to determine
the coordinates for the circle that will either capture all values, or those
within the quantile defined by \code{which_quant}. 
Multiple circles may be drawn based on groups provided via \code{separate_by}.
}
\details{
Sometimes, it is useful to remove outlier coordinates before drawing
the circles. If \code{keep_quant= 1}, the output will be exactly the content 
of \code{in.df}.
\code{keep_quant} should be a number between 1 and 0.5, usually you will want
to use something between 0.95 and 1.

Integer indeces for X and Y must be given in proper integer notation
(e.g., 1L) so that \code{is.integer(X)} passes.
}
\examples{
\dontrun{
c.dt <- fx.get_circle_df(df_to_plot, fill_by = circle_by, keep_quant= which_quantile)
plot_out <- plot_out +  ggalt::geom_encircle(data = c.dt,
                                             aes_string(fill = fx.parse_column_names(circle_by)),
                                             s_shape=0.9, expand=0.07, alpha = .1)
}

}
\seealso{
\code{\link{plot_reduced_dim.df}},
\code{\link{scABC2::generate.DimRedPlot}}
}
