% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dimRedPlot_fx.R
\name{fx.get_percent_var}
\alias{fx.get_percent_var}
\title{Extract the percent variation values stored as attributes}
\usage{
fx.get_percent_var(rd, pcs_select = NULL)
}
\arguments{
\item{rd}{PCA result object}

\item{pcs_select}{the number of PCs for which the values should be returned}
}
\description{
Extract the percent variation values stored as attributes
}
