% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dimRedPlot_fx.R
\name{fx.return_axis_length}
\alias{fx.return_axis_length}
\title{Get the number of coordinates stored in the axis value vector}
\usage{
fx.return_axis_length(axs)
}
\arguments{
\item{whatever}{is going to be passed on to get_reducedDimPlot.sce as the X
and Y coordinate vectors}
}
\value{
a number
}
\description{
The axis coordinates for tSNE and PCA can be given either as a
one-dimensional vector, or as a matrix or data.frame with one column. This
function returns the correct length that can be used to test against the 
number of cells stored in the singleCellExperiment object.
}
\examples{

df <- data.frame(a = c(1:10))
fx.return_axis_length(df[,1,drop=FALSE])
fx.return_axis_length(df[,1])

}
