% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drp_fx.R
\name{fx.set_factors}
\alias{fx.set_factors}
\title{Extract factor information}
\usage{
fx.set_factors(drp_object, fctr = color_by, fctr_type = "color_by",
  ignore_drp_labels)
}
\arguments{
\item{drp_object}{DimRedPlot object, ideally generated with \code{\link{}}}

\item{fctr}{the value of the given factor, e.g. "ENSMUSG00000025920" or 
"condition"}

\item{fctr_type}{the type of factor, i.e. the name such as "color_by", or
"size_by"}

\item{ignore_drp_labels}{either TRUE/FALSE or a vector of fctr_type(s) to be
ignored from the drp object}
}
\value{
the string that will be used to assign the fct
}
\description{
Extract factor information
}
