% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{gen.time}
\alias{gen.time}
\title{Function to calculate generation times (mean age of parents for a given cohort)
 NOTE: use a very large max age to reduce bias in numerator
 Liz Brooks
 Version 1.0
 Created 22 November 2013
 Last Modified:
------- Gen Time ----------------------------------}
\usage{
gen.time(nages, fec.age, mat.age, M.age, F.mult, sel.age, spawn.time)
}
\arguments{
\item{nages}{number of ages}

\item{fec.age}{fecundity at age}

\item{mat.age}{maturity at age}

\item{M.age}{natural mortality at age}

\item{F.mult}{F at fully selected ages}

\item{sel.age}{selectivity at age}

\item{spawn.time}{spawning time}
}
\description{
Function to calculate generation times (mean age of parents for a given cohort)
 NOTE: use a very large max age to reduce bias in numerator
 Liz Brooks
 Version 1.0
 Created 22 November 2013
 Last Modified:
------- Gen Time ----------------------------------
}
