% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genLogisticSelx.R
\name{genLogisticSelx}
\alias{genLogisticSelx}
\title{Logistic selectivity}
\usage{
genLogisticSelx(logisticA50, logisticSlope, ages)
}
\arguments{
\item{logisticA50}{age at 50 percent selectivity}

\item{logisticSlope}{slope at 50 percent selectivity}

\item{ages}{vector of ages to have selectivity calculated}
}
\description{
Create selectivity at age given parameters of a logistic function.
}
