% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{get.error.ac}
\alias{get.error.ac}
\title{Add autocorrelated error to a simulated process.}
\usage{
get.error.ac(ac, prev.err, evar)
}
\arguments{
\item{ac}{auto correlation coefficient}

\item{prev.err}{previous error e.g. result of a call to this function from a previous time step}

\item{evar}{the log space variance in your error term}
}
\description{
add an autocorrelated error term to a value method from Deroba & Bence 2012.
the result of this function can now be used as the error term in add.error.ac
}
