% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{get_fibonacci_n}
\alias{get_fibonacci_n}
\title{Get the nth element of the fibonacci vector}
\usage{
get_fibonacci_n(n)
}
\arguments{
\item{n}{Numeric, index of the vector to retrieve}
}
\value{
Numeric, nth element of the vector
}
\description{
Get the nth element of the fibonacci vector
}
\examples{
get_fibonacci_n(4)
}
