% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dimRedPlot_sce.R
\name{get_reducedDimPlot.sce}
\alias{get_reducedDimPlot.sce}
\title{Prepare the data.frame for plotting reduced dimension information}
\usage{
get_reducedDimPlot.sce(object, which_reddim = NULL, which_pcs = c(1, 2),
  X = NULL, Y = NULL, dim_red_type = NULL, color_by = NULL,
  shape_by = NULL, size_by = NULL, circle_by = NULL,
  exprs_values = "counts", add_cell_info = NULL)
}
\arguments{
\item{object}{\code{SingleCellExperiment} object}

\item{which_reddim}{accessor to retrieve the list of choice for the reduced
dimension coordinates from \code{reducedDims(object)}. See \code{details}.}

\item{which_pcs}{vector of integers to specify which (principal) components 
are going to be used for the x and y coordinates. The default (\code{c(1,2)})
will assign PC1 to the X axis and PC2 to the Y axis. This parameter will be
ignored if X and Y are specified directly.}

\item{X}{directly supply the X coordinates, e.g. \code{pca_res$PC1}; if this
is used, \code{which_reddim} and \code{which_pcs} are ignored.
\code{X} should be of the same length (and order!) as \code{colData(object)}
(or \code{pData2.0(object)}).}

\item{Y}{directly supply the Y coordinates, e.g. \code{pca_res$PC2}.
Should be of the same length (and order!) as \code{colData(object)} (or
\code{pData2.0(object)}).
\code{which_reddim} and \code{which_pcs} will be ignored if \code{X} and 
\code{Y} are specified.}

\item{dim_red_type}{what type of dimensionality reduction was used to generate
the data (e.g. "PCA"). Will mostly be used for labelling purposes. Default: NULL}

\item{color_by}{entry of either \code{colData(object)}, \code{rowData(object)},
\code{rownames(object)}, which will be used to assign either discrete or
continous color schemes. Alternatively, you can supply a \code{list} with a
label stored in \code{list$title} and values stored in \code{list$result}
(\code{result} should be as long as \code{dim(object)[2]}. Default: \code{NULL}}

\item{shape_by}{default: \code{NULL}}

\item{size_by}{default: \code{NULL}}

\item{circle_by}{Used for encircling points of interest; default: \code{NULL}}

\item{exprs_values}{types of expression values that will be used for plotting if a gene name is being
specified in either \code{color_by}, \code{shape_by}, \code{size_by}}

\item{add_cell_info}{if you want to add columns for certain entries of
\code{colData(object)}, specify their names here. Default: NULL}
}
\value{
DimRedPlot object: a simple S3-based object, i.e. a list with:
\enumerate{
\item \code{plot_data}: data.frame fit for ggplot2-based plotting
\item \code{label_data}: additional stuff that may be used for labelling the
x and y axes of the final plot
\item \code{factors}: names of the plotting variables, e.g. which values
were used for defining the coloring etc.
}
}
\description{
Wrapper function for extracting the X and Y coordinates for the
plot as well as the variables for coloring, shape assignment etc.
This function expects a SingleCellExperiment object.
}
\details{
\emph{which_reddim} The reducedDims slot of the SingleCellExperiment object
is a list (\code{\link{SimpleList}}. To know which coordinates to extract,
the accessor for the specific reducedDims list member must be given.
}
\examples{

data("tiny_sce")

# use reducedDimensions stored within the tiny_sce
drp <- get_reducedDimPlot.sce(tiny_sce, which_reddim = "PCA",
                                          which_pcs = c(2:3), 
                                          color_by = names(colData(tiny_sce))[1],
                                          dim_red_type = "PCA" )
head(drp$plot_data)
head(drp$label_data)

# specify X and Y manually
x <- reducedDim(tiny_sce, "PCA")[,"PC1", drop = FALSE]
y <- reducedDim(tiny_sce, "PCA")[,"PC2", drop = FALSE]

drp <- get_reducedDimPlot.sce(tiny_sce, X = x, Y = y)

## note the difference in the entries for drp$label_data depending on whether
## drop = FALSE is used to define x and y (if the colnames are dropped, they
## won't be part of the DimRedPlot object)

}
\seealso{
\code{\link{SingleCellExperiment}}, \code{\link{reducedDims}},
\code{\link{plt.DimRedPlot}}
}
