% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RuleBook.R
\name{get_RuleBook}
\alias{get_RuleBook}
\title{Get one of a series of predefined rule books for a model run.}
\usage{
get_RuleBook(book = "empty")
}
\arguments{
\item{book}{\code{Character} scalar, name of the rule book to be generated.
One out of \code{"empty"}, default is \code{"empty"}.}
}
\value{
A list object with all rules for a model run.
}
\description{
The function returns a prebuilt model rule book, i.e., a combination of
model parameters and rules.
}
\examples{

## create simple true age-depth-relationship
book_flat <- get_RuleBook(book = "empty")

}
\author{
Michael Dietze
}
