% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{getTimeFlag}
\alias{getTimeFlag}
\title{get a precise time flag from system time.}
\usage{
getTimeFlag()
}
\description{
get a precise time flag from system time.
}
\examples{
myflag=getTimeFlag()
print(myflag)
}
