\name{git_wf}
\alias{git_wf}
\title{Get github stats (watchers and forks).}
\usage{
  git_wf(id = "hadley", type = "user")
}
\arguments{
  \item{id}{name of the github user or organization}

  \item{type}{either "user" or "org"}
}
\description{
  This function gets watchers and forks of github repos for
  a user or organization.
}
\author{
  Scott Chamberlain
}

