% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_equal.R
\name{\%==\%}
\alias{\%==\%}
\title{Approximate Equality}
\usage{
a \%==\% b
}
\arguments{
\item{a}{Numeric vector.}

\item{b}{Numeric vector.}
}
\description{
Test for approximate equality of two vectors.
}
