% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{hello_world}
\alias{hello_world}
\title{hello world!}
\usage{
hello_world(n_exclamation)
}
\arguments{
\item{n_exclamation}{Numeric, integer indicating how many exclamation marks
to use}
}
\value{

}
\description{
hello world!
}
\examples{
hello_world(2)
hello_world(10)
}
