% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hl_chisq.R
\name{hl_chisq}
\alias{hl_chisq}
\title{Hosmer-Lemeshow Test}
\usage{
hl_chisq(obs, exp, ng = 10, lower.tail = FALSE)
}
\arguments{
\item{obs}{Observed counts}

\item{exp}{Expected counts}

\item{ng}{Number of groups (default is 10)}

\item{lower.tail}{Logical value for whether p-value should be from lower tail of the distribution.}
}
\description{
Tests for differences between observed and expected counts using the Hosmer-Lemeshow chi-squared test.
}
