% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe_chisq.R
\name{hwe_chisq}
\alias{hwe_chisq}
\title{Hardy-Weinberg Equilibrium Test}
\usage{
hwe_chisq(counts)
}
\arguments{
\item{counts}{Table of counts}
}
\description{
Tests for HWE for a table of counts of (aa, Aa, AA)
}
