% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe_matrix.R
\name{hwe_chisq_matrix}
\alias{hwe_chisq_matrix}
\title{Hardy-Weinberg Equilibrium Test}
\usage{
hwe_chisq_matrix(X)
}
\arguments{
\item{X}{Matrix of counts}
}
\description{
Tests for HWE for a matrix of table counts of (aa, Aa, AA)
}
