% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestROCthreshold.R
\name{indexOfUnion}
\alias{indexOfUnion}
\title{Index of Union}
\usage{
indexOfUnion(sens, spec, thr, auc)
}
\arguments{
\item{sens}{Sensitivities}

\item{spec}{Specificities}

\item{thr}{Threshold values for ROC}

\item{auc}{Area under the ROC curve}
}
\value{
A list containing the Index of Union value
}
\description{
This function will compute the Index of Union statistic
    which represents the ideal ROC threshold according to maximizing
    the statistic (sensitivity - auc) + (specificity - auc)
}
\examples{
# Simulate some data
sens <- runif(100,0,1)
spec <- runif(100,0,1)
thr <- runif(100,0,1)
auc <- runif(1,0,1)

indexOfUnion(sens, spec, thr, auc)

}
