% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infinite_loop.R
\name{infinite_loop}
\alias{infinite_loop}
\title{Run an infinite loop (for no obvious reason)}
\usage{
infinite_loop()
}
\description{
Run an infinite loop (for no obvious reason)
}
\examples{
infinite_loop()

}
