% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\alias{inv.logit}
\title{Logit and Inverse-Logit}
\usage{
inv.logit(x)

logit(p)
}
\arguments{
\item{x}{Vector of log-odds.}

\item{p}{Vector of probabilities.}
}
\description{
Logit and Inverse-logit transformations.
}
