% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_Sample.R
\name{make_Sample}
\alias{make_Sample}
\title{Create a virtual sample.}
\usage{
make_Sample(book, depth, geometry = "cuboid", radius, height, width, length)
}
\arguments{
\item{book}{\code{list} object, RuleBook to be used.}

\item{depth}{\code{Numeric} scalar, depth of the sample center (m).}

\item{geometry}{\code{Character} scalar, keyword defining the geometry of
the sample. One out of \code{"cuboid"} and \code{"cylinder"},
default is \code{"cuboid"}.}

\item{radius}{\code{Numeric} scalar, radius of the cylinder (m).}

\item{height}{\code{Numeric} scalar, height of the cuboid (m).}

\item{width}{\code{Numeric} scalar, width of the cuboid (m).}

\item{length}{\code{Numeric} scalar, length of the cuboid or cylinder (m).}
}
\value{
A list object.
}
\description{
The function generates many virtual sediment grains based on the specified
sample geometry and depth, using the information from a rule book.
}
\examples{

## To be done

}
\author{
Michael Dietze
}
