% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSandbox.R
\name{makeExternalRSession}
\alias{makeExternalRSession}
\title{Generate new R session and prep it to recieve a function}
\usage{
makeExternalRSession(sbConnection)
}
\arguments{
\item{sbConnection}{the sandbox connection object}
}
\description{
Actual function that generates the connection with the sandbox R session
}
