% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{MakeLetter}
\alias{MakeLetter}
\title{Put a letter in a plot (used for labeling a panel in a group of several plots)}
\usage{
MakeLetter(a = "A", where = "topleft", cex = 2)
}
\arguments{
\item{a}{default="A" the letter to put}

\item{where}{default="topleft"}

\item{cex}{default=2}
}
\description{
Put a letter in a plot (used for labeling a panel in a group of several plots)
}
