% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matts-colours.R
\docType{data}
\name{matt}
\alias{matt}
\alias{scale_color_matt}
\alias{scale_fill_matt}
\alias{matts_cols}
\alias{matts_pals}
\alias{matts_colours}
\alias{matts_palettes}
\title{Matts Custom Colours}
\format{An object of class \code{character} of length 3.}
\usage{
scale_color_matt(palette = "rainbow", discrete = TRUE, reverse = FALSE, ...)

scale_fill_matt(palette = "rainbow", discrete = TRUE, reverse = FALSE, ...)

matts_cols(...)

matts_pals(palette = "rainbow", reverse = FALSE, ...)

matts_colours

matts_palettes
}
\arguments{
\item{palette}{Name of palette.}

\item{discrete}{Logical. Is the colour palette for a discrete mapping?}

\item{reverse}{Logical. Should the colours/palette be reverse?}

\item{...}{Additional arguments passed on to relevant function.}
}
\description{
Functions for building and accessing Matt's favourite colours and palettes for use in ggplot visualizations.
}
\details{
Here is a breakdown of the important functions and what they do:

\itemize{
 \item \code{scale_color_matt} a ggplot-type function for providing colours for coloring a mapping.
 \item \code{scale_fill_matt} a ggplot-type function for providing colours for filling a mapping.
 \item \code{matts_colours} is a character vector of my favourite colours.
 \item \code{matts_palettes} is a list of curated palette sets.
 \item \code{matts_cols} is a function for access colours in \code{matts_colours} by name (hex value is returned).
 \item \code{matts_pal} returns a function to interpolate a colour palette from \code{matts_palettes}.
}

See \code{matts_palettes()} for a list of the available palettes.
}
\keyword{datasets}
