% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeData.R
\name{mergeSPGMICRSP}
\alias{mergeSPGMICRSP}
\title{Merge SPGMI scores with CRSP stocks}
\usage{
mergeSPGMICRSP()
}
\arguments{
\item{stocksCRSP}{a data.frame contains stocks in CRSP format. Default = 
stocksCRSP. See \link{stocksCRSP}}

\item{scoresSPGMI}{a data.frame contains factors in SPGMI format. Default =
scoresSPGMI. See \link{scoresSPGMI}}
}
\description{
Merge SPGMI scores with CRSP stocks
}
\examples{
data(stocksCRSP, package = "FactorAnalytics")
data(scoresSPGMI, package = "FactorAnalytics")
dataSPGMICRSP = mergeSPGMICRSP()
}
\author{
Kirk Li.
}
