% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{minmax}
\alias{minmax}
\title{Minimum-Maximum of Object}
\usage{
minmax(..., na.rm = TRUE)
}
\arguments{
\item{...}{Objects to determine min-max}

\item{na.rm}{Should NA be removed?}
}
\description{
Function to get min and max for a bunch of objects (all coerced to vectors) to set xlim and ylim in plot
}
