% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_utils.R
\name{my_gene}
\alias{my_gene}
\title{Map between ENSEMBL IDs and gene symbols}
\usage{
my_gene(goi, gn_map)
}
\arguments{
\item{goi}{vector of the IDs of the gene(s) of interest, e.g. "Ins2"}

\item{gn_map}{object that contains the mapping between gene_symbol and ensembl_id,
either an SCE object (from which \code{rowData()} will be used) or a data.table
with "gene_symbol" and "ensembl_id"}
}
\value{
either the gene_symbols (if ENSEMBL IDs were given) or the ENSEMBL IDs
(if gene_symbols were given) of the genes of interest
}
\description{
Convenience function for quickly extracting the gene symbol for
an ENSEMBL ID of interest or vice versa. This is needed often for the Lo data
set because the row names of the SCE object are ENSEMBL IDs.
}
\examples{
\dontrun{
gns <- rowData(sce.filt)[c(1:2)] \%>\% as.data.table
my_gene("Ins2", gns)
# [1] "ENSMUSG00000000215"
}

}
