% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nri.R
\name{nri}
\alias{nri}
\title{Net reclassification index (NRI)}
\usage{
nri(data, outcome, class1, class2)
}
\arguments{
\item{data}{Data}

\item{outcome}{Outcome variable}

\item{class1}{Classification system 1}

\item{class2}{Classification system 1}
}
\description{
Net reclassification index (NRI)
}
