% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{objectToString}
\alias{objectToString}
\title{Serialize Object to a string to transmit over socket}
\usage{
objectToString(object)
}
\arguments{
\item{object}{The object to be serialized}
}
\value{
serialized object string
}
\description{
Serialize Object to a string to transmit over socket
}
