% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offdiag.R
\name{offdiags}
\alias{offdiags}
\title{Offdiagnoal for Symmetric Matrix}
\usage{
offdiags(matrix, upper = T, both = F)
}
\arguments{
\item{matrix}{Matrix}

\item{upper}{Upper triangular portion}

\item{both}{Both upper and lower?}
}
\description{
Function to extract off-diagonal elements of a square matrix.
}
