% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{on_check}
\alias{on_check}
\title{Checkbox Change Event Handler}
\usage{
on_check(tag, msg)
}
\arguments{
\item{tag}{An htmltools tag.}

\item{msg}{Character string specifying the message type.}
}
\value{
The modified tag.
}
\description{
Dispatches the element's checked state (boolean) as a message.
}
\examples{
shiny::tags$input(type = "checkbox") |>
  on_check("toggle_agree")

}
