% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{on_click}
\alias{on_click}
\title{Click Event Handler}
\usage{
on_click(tag, msg, value = NULL)
}
\arguments{
\item{tag}{An htmltools tag.}

\item{msg}{Character string specifying the message type.}

\item{value}{Optional payload. An R value (serialized to JSON) or
a JavaScript expression string for a dynamic value.}
}
\value{
The modified tag.
}
\description{
Dispatches a message when the element is clicked.
}
\examples{
shiny::tags$button("+") |> on_click("increment")
shiny::tags$button("Page 2") |> on_click("set_page", 2)

}
