% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_roc.R
\name{optimal_roc}
\alias{optimal_roc}
\alias{index_of_union}
\alias{youden}
\alias{top_left}
\alias{min_pval}
\alias{conc_prob}
\title{Optimal ROC Threshold}
\usage{
index_of_union(sens = NULL, spec = NULL, thr = NULL, auc = NULL)

youden(sens = NULL, spec = NULL, thr = NULL)

top_left(sens = NULL, spec = NULL, thr = NULL)

min_pval(time = NULL, status = NULL, pred = NULL)

conc_prob(sens = NULL, spec = NULL, thr = NULL)
}
\arguments{
\item{sens}{Sensitivities}

\item{spec}{Specificities}

\item{thr}{Threshold values for ROC}

\item{auc}{Area under the ROC curve}

\item{time}{Time}

\item{status}{Censoring status}

\item{pred}{Predictions}
}
\value{
A scalar containing the chosen threshold.
}
\description{
Functions to identify the optimal ROC threshold according to various optimization criteria.
}
\details{
\itemize{

\item \strong{Youden's J Index}: This function will compute Youden's J index. This is computed by finding the threshold which maximizes the value \code{(sensitivity + specificity - 1)}

\item \strong{Index of Union}: This function will compute the Index of Union statistic which represents the ideal ROC threshold according to maximizing the statistic (sensitivity - auc) + (specificity - auc)

\item \strong{Top Left}: Provides the receiver operating characteristic (ROC) threshold
    (x-axis) that minimizes the distance from the ROC curve to the  to the
    top left corner of the plot (1,1)

\item \strong{Minimum P-value}: Provides the receiver operating characteristic (ROC) threshold
    that minimizes the P-value according to Unal (2017), Article ID 3762651

\item \strong{Concordance Probability}: Provides the receiver operating characteristic (ROC) threshold
    that maximizes the concordance probability. The concordance probability
    is the product of the sensitivity and specificity.

}
}
\examples{
# Simulate some data
sens <- runif(100,0,1)
spec <- runif(100,0,1)
thr <- runif(100,0,1)
auc <- runif(1,0,1)

index_of_union(sens, spec, thr, auc)
youden(sens, spec, thr)
top_left(sens, spec, thr)
conc_prob(sens, spec, thr)

}
