% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial-auc-spline.R
\name{partAUC}
\alias{partAUC}
\title{Partial AUC}
\usage{
partAUC(data = NULL, sens = NULL, spec = NULL, range = c(1, 0),
  focus = "sp", smooth = FALSE, correct = FALSE, n = 1000, ...)
}
\arguments{
\item{data}{Data frame.}

\item{sens}{Column name for sensitivity.}

\item{spec}{Column name for specificity.}

\item{range}{Range of sensitivity or specificity to integrate.}

\item{focus}{String describing whether the integration range is for sensitivity or specificity.}

\item{n}{Number of equally-spaced points for smooth curve calculations.}

\item{...}{Not currently used.}
}
\value{
\code{Tibble} containing a column for the partial AUC, and a list-column containing the ROC data.
}
\description{
Function for estimating the partial area under the ROC curve. See details for a more comprehensive description.
}
\details{
Since the x-axis of the ROC curve is mapped as 1 - Specificity, if the \code{focus} is "sp", the range should be provided in the left-to-right orientation for specificity. In other words, if you want to integrate the entire ROC in terms of specificity, the range would be c(1,0); this translates to integrating the false-positive rate from c(0,1).

If you only want to integrate the a section of the ROC curve (e.g. FPR from 0 to 0.2), you would specify the range in terms of specificity as c(1, 0.8).

When the focus is "se" for sensitivity, the range should be specified from the bottom-to-top (standard y-axis) orientation. In other words, to integrate the whole ROC curve, if the focus is "se" the range should be specificed as c(0,1). To integrate only a section of the ROC, for example, when sensitivity is between 0 and 0.5, you would specify the focus as "se" and the range as c(0,0.5).
}
\examples{
dd <- tibble::tibble(sens = runif(100), spec = runif(100))

partAUC(dd, sens, spec, range = c(1, 0.8))
partAUC(dd, sens, spec, range = c(1, 0.8), smooth = T, correct = T)

}
