% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Principal Component Analysis}
\usage{
pca(data, center = TRUE, scale = TRUE)
}
\arguments{
\item{data}{Data frame.}

\item{center}{Logical. Should the data be centered?}

\item{scale}{Logical. Should the data be scaled by the standard deviation?}
}
\value{
Tibble projected on the new PCA space.
}
\description{

}
\details{
Perform principal component anlaysis on a numeric data frame.
}
