% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink-scoring.R
\name{plink_scoring}
\alias{plink_scoring}
\title{PLINK Allelic Scoring}
\usage{
plink_scoring(bfile, scores, sum = TRUE, header = TRUE,
  cols = "1 2 3", range_file = "", data_file = "", ...)
}
\arguments{
\item{bfile}{Path to \code{PLINK} files (bed, bim, fam); prefix only.}

\item{scores}{Path to scoring file; see details.}

\item{sum}{Logical. Should sum of allele scores be returned? Average allele scores returned if set to \code{FALSE}.}

\item{header}{Logical. Should \code{PLINK} ignore the first non-empty line (assumed to be a header) in \code{scores} and \code{data_file} (if specified)?}

\item{cols}{A string specifying the column positions for the variant names, allele codes, and scores. By default, \code{PLINK} assumes the column positions are 1, 2, and 3 in \code{scores}, respectively.}

\item{range_file}{Path to file containing range labels in the first column (i.e. named ID for a given range), lower bounds in the second column, and upper bounds in the third column. See details.}

\item{data_file}{Path to file containing variant IDs (column 1) and the key quantity (column 2) on each non-empty line. See details.}

\item{...}{Additional arguments passed to \code{\link[readr]{read_delim}}.}
}
\value{
A \code{\link[tibble]{tibble}} with six columns:
\itemize{
  \item \strong{\code{FID}} Family ID (from \code{PLINK} .fam file)
  \item \strong{\code{IID}} Individual ID (from \code{PLINK} .fam file)
  \item \strong{\code{PHENO}} Phenotype (from \code{PLINK} .fam file)
  \item \strong{\code{CNT}} Number of non-missing alleles used for scoring
  \item \strong{\code{CNT2}} Sum of named allele counts
  \item \strong{\code{SCORE/SCORESUM}} Either the average of all allele scores, or the sum of the allele scores, depending on the \code{sum} argument
}

If allelic scoring is done for multiple subsets of the variants, defined by \code{range_file} and \code{data_file}, then each plink.*.profile file is read, row-bound, and nested by file-name, to form a single nested \code{\link[tibble]{tibble}}. The \code{nested-tibble} contains a column for the file name and a list-column of \code{tibbles} for the nested scoring data. Use \code{\link[tidyr]{unnest}} to unnest the data back into a single \code{tibble}.

The returned \code{\link[tibble]{tibble}} has the added attribute \code{log} which contains a \code{\link[tibble]{tibble}} for the log-file from \code{PLINK}. This can be accessed using \code{attr(x, 'log')}, where \code{x} is the name of your object.

The \code{log} attribute contains a one-column \code{\link[tibble]{tibble}} with a row for every line-break in the log-file. This allows relatively easy access to the log while staying in \code{R}, and one can utilize \code{\link{stringr}} functions to query the log messages.
}
\description{
Interface to \href{http://www.cog-genomics.org/plink/1.9/score}{PLINK} to perform allelic scoring. This function will save the \code{PLINK} outputs to the current working directory, and read plink.profile into \code{R} for immediate use as a \code{\link[tibble]{tibble}}.
}
\details{
The \code{scores} file is white-space delimited, and should contain a row for every variant included in the scoring algorithm, and should have a minimum of 3 columns:
\enumerate{
  \item Variant name (usually rsID)
  \item Allele code (the allele that the score is in reference too)
  \item Score (score associated with the named allele)
}

In other words, for a given variant (column 1), the score (column 3) may represent the per-allele increase in the log-odds of the phenotype for every additional allele named in column 2 (i.e. additive genetic model).

\code{PLINK} also offers the ability to apply allelic scoring to a subset(s) of the variants in \code{scores} based on the range of some key quantity (e.g. P-value). To do this, you must additionally provide:

\itemize{
  \item \strong{\code{range_file}} A white-space delimited text file with three columns (header optional; \code{PLINK} ignores non-numeric values in columns 2 and 3, by default). The first column is a unique ID for the range specified in columns 2 (lower bound) and 3 (upper bound), inclusive. This ID is used in producing the named output from \code{PLINK} (e.g. plink.ID.profile)
  \item \strong{\code{data_file}} A white-space delimited text file with two columns (header will be ignored or not based on the \code{header = TRUE/FALSE} argument; so presence of a header must be consistent with the \code{scores} file). Column 1 is the variant ID and column 2 is the key quantity that the range will be applied to (e.g. P-value)
}

For example, a range of [0, 0.00000005] would perform allelic scoring for all variants in \code{scores} and \code{data_file} that have a GWAS-significant P-value.
}
