% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portWeights.R
\name{portWeights}
\alias{portWeights}
\title{portWeights}
\usage{
portWeights(returns)
}
\arguments{
\item{returns}{A K-dimensional xts time series of returns}
}
\value{
A K-dimensional xts time series of portfolio weights
}
\description{
Creates portfolio weights for a specified portfolio. In
this version it just computes a global minimum variance long-only
portfolio (gmvLO).
}
\details{
This function use the PortfolioAnalytics, ROI,
and ROI.plugin.quadprog packages
}
\examples{
library(FactorAnalytics)
library(xts)
data(stocksCRSP)
data(factorsSPGMI)
stocks <- selectCRSPandSPGMI(stocksCRSP,factorsSPGMI)
returnMat = tapply(stocks[["Return"]],list(stocks$Date,stocks$Ticker),I)
returns = xts(returnMat,as.yearmon(rownames(returnMat)))
wtsGmvLO <- portWeights(returns)
}
\references{
Martin et al. (2022) Portfolio Construction and Risk Analysis, Chapter 4.
}
\author{
Doug Martin
}
