% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_train_test.R
\name{pred_train_test}
\alias{pred_train_test}
\title{Training and Testing Predictions}
\usage{
pred_train_test(model, train, test, type = "response")
}
\arguments{
\item{model}{Model object}

\item{train}{Training data}

\item{test}{Testing data}

\item{type}{Type of prediction; passed to \code{predict}}
}
\description{
Make predictions using in \code{train} and \code{test} data using a \code{model}. The \code{model} object must have a S3 predict method. In other words, there should be a \code{predict.class} for whatever class the model is. To check the class of a model run \code{class model_obj}. Also, check \code{methods('predict')} to check if your model class is supported as a S3 predict method.
}
\details{
Both training and testing must contian all the predictors used in the model. Predictor names must be a full match.
}
