% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predBevHoltSR.R
\name{predBevHoltSR}
\alias{predBevHoltSR}
\title{Beverton Holt Stock Recruitment}
\usage{
predBevHoltSR(BHalpha, BHbeta, SSB)
}
\arguments{
\item{BHalpha}{Bev-Holt alpha parameter in R = alpha * SSB / (beta + SSB)}

\item{BHbeta}{Bev-Holt beta parameter in R = alpha * SSB / (beta + SSB)}

\item{SSB}{one or more SSB values}
}
\description{
Calculates expected value from BH SR curve for given SSB.
}
