% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_Aliquot.R
\name{prepare_Aliquot}
\alias{prepare_Aliquot}
\title{Prepare aliquots from sample data set}
\usage{
prepare_Aliquot(sample, diameter, density = 0.65)
}
\arguments{
\item{sample}{\code{Data frame}, sample object to be distributed to 
aliquots.}

\item{diameter}{\code{Numeric} value, diameter of the alquot sample 
carriers in mm.}

\item{density}{\code{Numeric} value, packing density of the grains on
the sample carrier. Default is \code{0.65}.}
}
\value{
\code{List} object with grains organised as aliquots, i.e. list 
elements.
}
\description{
The function consecutively fills aliquots (i.e., subsamples distributed on 
round carrier discs) with grains from an input sample. remaining grains that 
are not enough to fill a further aliquot are discarded.
}
\examples{
## load example data set
data(sample)

A <- prepare_Aliquot(sample = sample, 
                     diameter = 5)

B <- prepare_Aliquot(sample = sample, 
                     diameter = 2, 
                     density = 0.6)

}
\author{
Michael Dietze
}
