% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox_object.R
\name{print.sandbox.output}
\alias{print.sandbox.output}
\title{print method for sandbox.output}
\usage{
\method{print}{sandbox.output}(x, env = parent.frame())
}
\arguments{
\item{x}{sandbox.output to be printed}
}
\description{
print method for sandbox.output
}
