% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox_object.R
\name{print.sandbox_output}
\alias{print.sandbox_output}
\title{print method for sandbox.output}
\usage{
\method{print}{sandbox_output}(x, ..., env = parent.frame())
}
\arguments{
\item{x}{sandbox.output to be printed}

\item{...}{additional arguments passed to print}

\item{env}{environment to assign leaked objects to}
}
\description{
print method for sandbox.output
}
