% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pz-converter.R
\name{pz}
\alias{pz}
\alias{z_to_p}
\alias{p_to_z}
\title{P-value and Z-Score Converter}
\usage{
z_to_p(z)

p_to_z(p, sign)
}
\arguments{
\item{z}{Z-score.}

\item{p}{P-value.}

\item{sign}{Sign (-1/+1) of beta.}
}
\value{
Either a P-value or Z-score.
}
\description{
Two simple functions to convert between P-values and Z-scores.
}
