% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socketCommunication.R
\name{receiveSand}
\alias{receiveSand}
\title{share and send objects over sockets
Get code to evaluate from original R Session}
\usage{
receiveSand(sandboxCon)
}
\arguments{
\item{sandboxCon}{socket connection to original R session}
}
\description{
share and send objects over sockets
Get code to evaluate from original R Session
}
