% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-beamer.R
\name{render_beamer}
\alias{render_beamer}
\title{Render R Markdown to Beamer Presentation}
\usage{
render_beamer(input = NULL, cleanup = FALSE, ...)
}
\arguments{
\item{input}{R Markdown File. See details.}

\item{cleanup}{Logical. Should the intermediate \code{TeX} files be removed? Default is FALSE.}

\item{...}{Additional named arguments passed to \code{\link[rmarkdown]{render}}.}
}
\value{
Invisibly, the rendered PDF path and file name.
}
\description{
Render R Markdown files destined to be Beamer PDF presentations with a double-rendering approach using a \code{LaTeX} engine of your choosing (or \code{pdflatex}, by default), to produce proper slide numbering.
}
\details{
This function is a crude solution to the issue whereby R Markdown (\code{.Rmd}) documents being rendered to Beamer PDF presentations fail to render the total number of slides properly. This problem did not exist in previous verions, but currently exists.

Basically, this function renders the R Markdown to Markdown to \code{TeX} with a call to \code{\link[rmarkdown]{render}}. Then, the function uses a \code{LaTeX} engined specified by \code{options(latex.engine = "your/path/here")} or tries \code{pdflatex}, by default. The \code{TeX} file is rendered twice and then opened for viewing.

When specifying the \code{input} for this function, the path included in the input is used for writing any new files (including the final PDF). If only a file name is provided with no path, it is assumed that the current working directory should be used for writing.
}
