% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder-terms.R
\name{reorder_terms}
\alias{reorder_terms}
\title{Reorder terms for kable table}
\usage{
reorder_terms(data, ..., order = letters, rev = FALSE)
}
\arguments{
\item{data}{Data frame.}

\item{...}{Column names to be ordered.}

\item{order}{Character vector of term names, in the desired order.}

\item{rev}{Logical. Should the order be reversed?}
}
\value{
A \code{\link[tibble]{tibble}} with arranged columns.
}
\description{
Function to order the terms (usually regression variables) in a specific order to present in your table.
}
