% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{resize.win}
\alias{resize.win}
\title{adjust window size regardless of platform}
\usage{
resize.win(Width = 10, Height = 10, xpinch = 20, ypinch = 30)
}
\arguments{
\item{Width}{default=10 inches}

\item{Height}{default=10 inches}

\item{xpinch}{pixels per inch default=20}

\item{ypinch}{pixels per inch default=30}
}
\description{
adjust window size regardless of platform
}
