% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_Utils.R
\name{returnSand}
\alias{returnSand}
\title{Return outputs to original R Session - wrapper for sendSand}
\usage{
returnSand(sandboxCon, results)
}
\arguments{
\item{sandboxCon}{socket connection to original R session}

\item{results}{results from evaluated code}
}
\description{
Return outputs to original R Session - wrapper for sendSand
}
