% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox-package.R
\docType{package}
\name{sandbox-package}
\alias{sandbox}
\alias{sandbox-package}
\title{sandbox: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Joyce Robbins \email{jtr13@columbia.edu}

}
\keyword{internal}
