% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox_main_help_page.R
\docType{package}
\name{sandbox}
\alias{sandbox}
\alias{sandbox-package}
\title{Test Stuff Out}
\description{
This is just a sandbox for me to test out package stuff and Github.
}
\section{Notes}{

This help page should be in synce with DESCRIPTION. Easy for users to find
using \code{?<package-name>}. Should describe the most important components of the
package.
}

