% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox_handle.R
\name{sandbox_handler}
\alias{sandbox_handler}
\title{sandbox_evaluate_handler}
\usage{
sandbox_handler()
}
\description{
expand upon basic output_handler from evaluate. Really only adds returning the actual value to return
}
