% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandboxConnectionTemplate.R
\name{sandboxConnectionTemplate}
\alias{sandboxConnectionTemplate}
\title{Contains template information for connection to external server}
\usage{
sandboxConnectionTemplate(host = "localhost", port, username = "",
  hostOS = c("unix", "windows"))
}
\arguments{
\item{host}{server where  R code will be exectuted, defaults to local}

\item{port}{port of the connection}

\item{username}{external servers username}

\item{hostOS}{the OS of the external host, unix or windows?}
}
\description{
Contains template information for connection to external server
}
