% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socketCommunication.R
\name{sendMessage}
\alias{sendMessage}
\title{fascilitate communication, send message with unique ID. wait for confirmation of recept of message with unique identifier confirmation}
\usage{
sendMessage(con, string, maxlen = 256L)
}
\description{
fascilitate communication, send message with unique ID. wait for confirmation of recept of message with unique identifier confirmation
}
