% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socketCommunication.R
\name{sendSand}
\alias{sendSand}
\title{share and send objects over sockets
Get code to evaluate from original R Session}
\usage{
sendSand(expr, sandboxCon)
}
\arguments{
\item{expr}{the expression to be sent to the r session on the other end of the `sandboxCon`}

\item{sandboxCon}{socket connection to original R session}
}
\description{
share and send objects over sockets
Get code to evaluate from original R Session
}
