% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Parameter.R
\name{set_Parameter}
\alias{set_Parameter}
\title{Set profile- and grain-specific model parameters.}
\usage{
set_Parameter(book, parameter, type)
}
\arguments{
\item{book}{\code{list} object, rule book to be edited.}

\item{parameter}{\code{Character} scalar, keyword defining the parameter to be
defined. See \code{list_Parameters()} for available keywords. Some
parameters can be described by more than one function, see details.}

\item{type}{\code{Character} scalar, keyword defining the distribution
function used to describe the parameter. See details for available
keywords, default is "constant".}
}
\value{
A list object.
}
\description{
The function defines one model parameter used to generate a set of
virtual grains. A parameter is defined in a probabilistic way, as parametric
distribution function. Each parameter of the distribution function can be
changed through time using \code{set_Rule}.
}
\examples{

## get empty rule book
book_1 <- get_RuleBook(book = "empty")

## set dose rate parameterisation from default to "rnorm"
book_2 <- set_Parameter(book = book_1,
                        parameter = "dose_rate",
                        type = "rnorm")

book_1$dose_rate
book_2$dose_rate

## set grainsize parameterisation from default to "constant"
book_3 <- set_Parameter(book = book_1,
                        parameter = "grainsize",
                        type = "constant")

book_1$grainsize
book_3$grainsize

}
\author{
Michael Dietze
}
