% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Rule.R
\name{set_Rule}
\alias{set_Rule}
\title{Set depth-dependent rule for model parameter.}
\usage{
set_Rule(book, parameter, value, depth, type = "spline")
}
\arguments{
\item{book}{\code{List} object, rule book to be edited.}

\item{parameter}{\code{Character} scalar, parameter name to be edited.}

\item{value}{\code{Numeric} list, specifying the
parameter values at the corresponding depth points. If a parameter
is defined by more than one argument (e.g., mean and standard
deviation), all the relevant arguments must be defined for each
corresponding depth as separate list element.}

\item{depth}{\code{Numeric} list, specifying the depths used for the
interpolation. All elements must be of the same lengths as the
corresponding data in \code{value}.}

\item{type}{\code{Character} scalar, interpolation method. One out of
\code{spline}, default is \code{spline}.}
}
\value{
A list object with all created formula objects.
}
\description{
The function defines how the specified model parameter varies with depth.
The transfer function uses different interpolation functions to create a
continuous representation of a parameter value with depth.
}
\examples{

## create simple true age-depth-relationship
## TO BE DONE.


}
\author{
Michael Dietze
}
