% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-seed.R
\name{set_seed}
\alias{set_seed}
\title{Set Seed for RNG}
\usage{
set_seed()
}
\value{
Returns a message with the set-seed value.
}
\description{
User-friendly function for interactively setting the seed for random-number generation (RNG).
}
\examples{

set_seed()
1234

}
