% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymvu-package.R
\docType{package}
\name{shinymvu-package}
\alias{shinymvu}
\alias{shinymvu-package}
\title{shinymvu: The Elm Architecture for Shiny Applications}
\description{
Model-View-Update architecture for Shiny applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/armcn/shinymvu}
  \item Report bugs at \url{https://github.com/armcn/shinymvu/issues}
}

}
\author{
\strong{Maintainer}: Andrew McNeil \email{andrew.richard.mcneil@gmail.com}

}
