% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snptest.R
\name{snptest_sum}
\alias{snptest_sum}
\title{SNPTEST Functions}
\usage{
snptest_sum(data, sample, out, ...)
}
\arguments{
\item{data}{BGEN or PLINK format genotypes.}

\item{sample}{Sample file.}

\item{out}{Path and name of output file.}

\item{...}{Additional arguments.}

\item{freq}{Frequenstist metho}

\item{method}{method}

\item{pheno}{peh}
}
\description{
Functions for computing summary statistics or Frequentist Association Tests.
}
