% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{stringToObject}
\alias{stringToObject}
\title{Unserialize Object from a string to transmit over socket}
\usage{
stringToObject(string)
}
\arguments{
\item{string}{A string that once decompressed/unserialized is an object}
}
\value{
object that was serialized
}
\description{
Unserialize Object from a string to transmit over socket
}
