% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_first_fibonacci_numbers.R
\name{sum_n_first_fibonacci_numbers}
\alias{sum_n_first_fibonacci_numbers}
\title{sum the first n fibonacci numbers}
\usage{
sum_n_first_fibonacci_numbers(n)
}
\arguments{
\item{n}{Numeric, integer indicating how many numbers to sum}
}
\value{
the num of the n first fibonacci numbers
}
\description{
sum the first n fibonacci numbers
}
\examples{
sum_n_first_fibonacci_numbers(5)
sum_n_first_fibonacci_numbers(8)
}
