% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_square_fibonacci.R
\name{sum_square_fibo}
\alias{sum_square_fibo}
\title{Sum up the first nth fibonacci numbers with power 2.}
\usage{
sum_square_fibo(n)
}
\arguments{
\item{n}{numeric, positive integer}
}
\value{
print a prompt with "Sum of the first nth of Fibonacci numbers of power 2:",
along with the sum square value.
}
\description{
Sum up the first nth fibonacci numbers with power 2.
}
\examples{
sum_square_fibo(2)
sum_square_fibo(3)
}
