% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-one.R
\name{table1}
\alias{table1}
\alias{table1_numeric}
\alias{table1_categorical}
\title{Table One Tibbles}
\usage{
table1_numeric(data, ..., group = NULL, round = 2, digits = round,
  total = FALSE, na.rm = FALSE)

table1_categorical(data, ..., group = NULL, round = 2,
  digits = round, total = FALSE, na.rm = FALSE,
  na_level = "(Missing)")
}
\arguments{
\item{data}{Data frame.}

\item{...}{Columns to summarise.}

\item{group}{Grouping variable for group summaries.}

\item{round}{Number of digits for rounding statistics (where applicable).}

\item{digits}{Number of significant digits to print (pad with zeroes if digits < round)}

\item{total}{Logical. Should a total column be added? (Default is FALSE) This argument only matters when a grouping variable is provided.}

\item{na.rm}{Logical. Should missing data be remove before computing summary statisitcs? (Default is FALSE)}
}
\value{
A \code{\link[tibble]{tibble}} with as many rows as the number of columns provided to \code{...}, and as many columns as levels in \code{group}. Cell values are \code{\link[glue]{glue}} objects.
}
\description{
Functions for producing descriptive statistics for numeric and categorical data. The objective of these functions is to make it easier to produce "Table 1" for a manuscript.

Tables are seperately created for numeric and categorical variable, and need to be combined manually afterward (probably using \code{\link[dplyr]{bind_rows}}).
}
\details{
For numeric variables passed to \code{table1_numeric}, the returned summary statistics are means (\code{\link[base]{mean}}) and standard deviations (\code{\link[stats]{sd}}). For categorical variables (i.e. factors) passed to \code{table1_categorical}, the returned summary statistics are counts (\code{\link[base]{sum}}) and percentages (\code{\link[base]{mean}}) of indicator variables for each level in a factor variable.
}
