% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers-competing-risk.R
\name{tidy.crr}
\alias{tidy.crr}
\title{Broom tidiers for competing risk models.}
\usage{
tidy.crr()
}
\description{
`tidy`, `glance`, and `augment` functions for tidying up model results for competing risk analysis.
}
