% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-robust-poisson.R
\name{tidy_robpois}
\alias{tidy_robpois}
\title{Tidy Robust Poisson Regression}
\usage{
tidy_robpois(data, formula, ci = 0.95, expo = FALSE)
}
\arguments{
\item{data}{Data frame.}

\item{formula}{Formula describing the model.}

\item{ci}{Confidence interval width.}

\item{expo}{Logical. Should the estimates and confidence intervals be exponentiated?}
}
\value{
A \code{tibble}.
}
\description{
Fits a Poisson regression model with robust variance estimation (modified Poisson regressin) and returns a tidy version of the results.
}
