% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{time}
\alias{time}
\title{Display time graphically}
\usage{
time(military = FALSE)
}
\arguments{
\item{military}{Logical (default = FALSE)}
}
\description{
Display time graphically
}
\examples{
time()
time(military = TRUE)

}
