% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timebars.R
\name{timebars}
\alias{timebars}
\title{Display time graphically}
\usage{
timebars(military = FALSE)
}
\arguments{
\item{military}{Logical (default = FALSE)}
}
\description{
Display time graphically
}
\examples{
timebars()
timebars(military = TRUE)

}
