% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{tiny_sce}
\alias{tiny_sce}
\title{Tiny SingleCellExperiment object for testing purposes}
\format{A \code{SingleCellExperiment} object with 10 rows (genes) and 10 
columns (cells).
Most of the \code{colData} was added by scater's calculateQCmetrics function.}
\usage{
tiny_sce
}
\description{
SingleCellExperiment object that can be loaded for testing the
functions.
}
\seealso{
\code{\link{SingleCellExperiment}}
}
\keyword{datasets}
