% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestROCthreshold.R
\name{topleft}
\alias{topleft}
\title{Closest to Top Left}
\usage{
topleft(sens, spec, thr)
}
\arguments{
\item{sens}{Sensitivity}

\item{spec}{Specificity}

\item{thr}{Threshold}
}
\value{
A list with two elements. The first element is a character string
    letting you know which result is returned. The second element is the
    top-left threshold
}
\description{
Provides the receiver operating characteristic (ROC) threshold
    that maximizes the concordance probability. The concordance probability
    is the product of the sensitivity and specificity.
}
\examples{

sens <- runif(100, 0, 1)
spec <- runif(100, 0, 1)
thr <- runif(100, 0, 1)

topleft(sens, spec, thr)

}
