\name{totimp_hello}
\alias{totimp_hello}
\title{Check that Total Impact (http://totalimpact.org/) API works - hello world.}
\usage{
  totimp_hello(url = "http://total-impact-core.herokuapp.com/v1")
}
\arguments{
  \item{url}{The base URL (do not change from default).}
}
\value{
  Hello world message, including basic info on the API.
}
\description{
  Check that Total Impact (http://totalimpact.org/) API
  works - hello world.
}
\examples{
\dontrun{
totimp_hello()
}
}

