% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulFunctions.R
\name{weighted.median}
\alias{weighted.median}
\title{Compute a weighted median of a numeric vector.}
\usage{
weighted.median(x, w, na.rm = TRUE, ties = NULL)
}
\arguments{
\item{x}{a numeric vector containing the values whose weighted median is
to be computed.}

\item{w}{a vector of weights the same length as x giving the
weights to use for each element of x. Default value is equal
weight to all values.}

\item{na.rm}{a logical value indicating whether NA values in
x should be stripped before the computation proceeds.}

\item{ties}{a character string specifying how to solve ties between two
         x's that are satisfying the weighted median criteria.
         Note that at most two values can satisfy the criteria.
         When {ties} is {"min"}, the smaller value of the two
         is returned and when it is {"max"}, the larger value is
         returned.
         If {ties} is {"mean"}, the mean of the two values is
         returned and if it is {"both"}, both values are returned.
         Finally, if {ties} is {"weighted"} (or {NULL}) a
         weighted average of the two are returned, where the weights are
         weights of all values {x[i] <= x[k]} and {x[i] >= x[k]},
         respectively. Default value is {NULL}.
Returns the weighted median.}
}
\description{
Compute a weighted median of a numeric vector.
}
