% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{wrap}
\alias{wrap}
\title{wrap user solution in a function}
\usage{
wrap(expr, G)
}
\arguments{
\item{expr}{expression}

\item{G}{given constants}
}
\description{
wrap user solution in a function
}
\keyword{internal}
