% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeASAP.R
\name{writeASAP}
\alias{writeASAP}
\title{Write ASAP data file}
\usage{
writeASAP(om, ASAPoptions, ASAPinputFileName, wd)
}
\arguments{
\item{om}{operating model list containing data}

\item{ASAPoptions}{list of options to use in the ASAP input file}

\item{ASAPinputFileName}{name for ASAP input file}

\item{wd}{working directory to write ASAP input file to}
}
\description{
Writes an ASAP input based on operating model OM.
}
