% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestROCthreshold.R
\name{youden}
\alias{youden}
\title{Youden's J Index}
\usage{
youden(sens, spec, thr)
}
\arguments{
\item{sens}{Sensitivities}

\item{spec}{Specificities}

\item{thr}{Threshold values for ROC}
}
\value{
A list containing Youden's Index
}
\description{
This function will compute Youden's J index. This is computed
    by finding the threshold which maximizes the value
    \code{(sensitivity + specificity - 1)}
}
\examples{
# Simulate some data
sens <- runif(100,0,1)
spec <- runif(100,0,1)
thr <- runif(100,0,1)

youden(sens, spec, thr)

}
