% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_train_test.R
\name{auc_train_test}
\alias{auc_train_test}
\title{Area Under the Curve}
\usage{
auc_train_test(model, train, test, truth, type = "response")
}
\arguments{
\item{model}{Model object.}

\item{train}{Training data.}

\item{test}{Testing data.}

\item{truth}{Ground-truth label.}

\item{type}{Type of prediction; passed to predict.}
}
\description{
Get area under the ROC curve for both training and test data.
}
