% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Competing Risk Calibration}
\usage{
calibrate(data, horizon, groups = 10, cause = 1, model)
}
\arguments{
\item{data}{Dataframe to use for absolute risk predictions}

\item{horizon}{Time horizon for predictions}

\item{groups}{Number of deciles based on model-predicted risks (default is 10)}

\item{cause}{Integer value for the event-of-interest to calibrate predictions}

\item{model}{The returened model object from \code{\link[riskRegression]{riskRegression}}}
}
\value{
A list containing three elements. The first element returns a string
   detailing the time horizon for calibration. The second element is a chi-sq
   test statistic and p-value for the Hosmer-Lemeshow test (with DF=groups-1).
   The third element is a \code{tibble} containining the decile index
   along with the predicted and observed cumulative incidences, and the
   predicted and observed events across the \code{n} risk quantiles.
}
\description{
This function will allow you to compute calibration metrics:
   - predicted and observed probabilities
   - predicted and observed events of interest
}
\examples{
library(riskRegression)

# Simulate Data
sim.data <- sampleData(100, outcome = 'competing.risks')

# Fit Fine-Gray Model
fgr.mod <- FGR(formula = Hist(time, event) ~ X7 + X6 + X8 + X9,
               data = sim.data, cause = 1)

calibrate(sim.data, 2, 10, 1, fgr.mod)

}
\author{
Matthew T. Warkentin
}
