% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{evaluateSandbox}
\alias{evaluateSandbox}
\title{Make Sandbox Run}
\usage{
evaluateSandbox(ID)
}
\arguments{
\item{ID}{the port ID to open socket on}
}
\description{
wrapper function to assist with running in sandbox
}
