% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drp.R, R/prep_dimRedPlot_sce.R
\name{get_reducedDimPlot.sce}
\alias{get_reducedDimPlot.sce}
\alias{plt.DimRedPlot}
\alias{get_reducedDimPlot.sce}
\title{Visualizing tSNE and PCA results}
\usage{
plt.DimRedPlot(drp_object, X = "x_axs", Y = "y_axs", color_by = NULL,
  shape_by = NULL, size_by = NULL, circle_by = NULL, which_quantile = 1,
  ignore_drp_labels = FALSE, theme_size = 14, legend = "auto",
  alpha = 0.75, remove_rug = FALSE, set_colors = TRUE,
  set_fill_colors = TRUE)

get_reducedDimPlot.sce(object, which_reddim = NULL, which_pcs = c(1, 2),
  X = NULL, Y = NULL, dim_red_type = NULL, color_by = NULL,
  shape_by = NULL, size_by = NULL, circle_by = NULL,
  exprs_values = "counts", add_cell_info = NULL)
}
\arguments{
\item{drp_object}{a \code{\link{DimRedPlot}} object}

\item{X}{name of the drp_object column that contains the values for the x axis}

\item{Y}{name of the drp_object column that contains the values for the y axis}

\item{color_by}{specify the column of \code{drp_object$plot_data} whose
entries will be used to assign either discrete or continous color schemes.}

\item{shape_by}{the factor to assign the shape of the points to; max. 10 levels
are accepted for \code{shape_by}. If you want to specify the (same) shape for
\emph{all} points, use an integer (default shape is 19).}

\item{size_by}{the factor to assign the size of the points to;
default: \code{NULL}. If you want to specify the (same) size for \emph{all}
points, use one integer (the default size is 3).}

\item{circle_by}{a factor (e.g. "condition") that will be used to draw circles
around the points belonging to the same factor instance.}

\item{which_quantile}{default: 1}

\item{ignore_drp_labels}{If set to TRUE, \emph{all} entries of
\code{drp_object$factors} will be ignored. If you want to ignore only selected
labels, indicate their name (e.g. c("color_by", "exprs_val_type")).
The default setting (\code{FALSE}) will make use of the information from 
the drp_object \emph{unless} specified here via \code{color_by}, \code{shape_by},
\code{size_by}, \code{circle_by}.}

\item{theme_size}{font size for the plot}

\item{legend}{choose wether there should be no legend ("none"), a legend for 
every factor or whether the legend will only show up for those factors that
have more than 1 level ("auto", default setting).}

\item{alpha}{define the opacity of the points, default: .65}

\item{remove_rug}{Define whether the bars at both axes that
represent the two 1d marginal distributions should be removed. Default: FALSE.}

\item{set_colors}{Set to \code{FALSE} if you want to add your own color scheme
to the resulting plot. The default behavior is to try to automatically assign
an optimized coloring scheme. Default: \code{TRUE}.}

\item{set_fill_colors}{Set to \code{FALSE} if you want to add your own color
scheme to the resulting plot \emph{for the background circles}. The default
behavior is to try to automatically assign an optimized coloring scheme.}

\item{object}{\code{SingleCellExperiment} object}

\item{which_reddim}{accessor to retrieve the list of choice for the reduced
dimension coordinates from \code{reducedDims(object)}. See \code{details}.}

\item{which_pcs}{vector of integers to specify which (principal) components 
are going to be used for the x and y coordinates. The default (\code{c(1,2)})
will assign PC1 to the X axis and PC2 to the Y axis. This parameter will be
ignored if X and Y are specified directly.}

\item{dim_red_type}{what type of dimensionality reduction was used to generate
the data (e.g. "PCA"). Will mostly be used for labelling purposes. Default: NULL}

\item{exprs_values}{types of expression values that will be used for plotting if a gene name is being
specified in either \code{color_by}, \code{shape_by}, \code{size_by}}

\item{add_cell_info}{if you want to add columns for certain entries of
\code{colData(object)}, specify their names here. Default: NULL}

\item{X}{directly supply the X coordinates, e.g. \code{pca_res$PC1}; if this
is used, \code{which_reddim} and \code{which_pcs} are ignored.
\code{X} should be of the same length (and order!) as \code{colData(object)}
(or \code{pData2.0(object)}).}

\item{Y}{directly supply the Y coordinates, e.g. \code{pca_res$PC2}.
Should be of the same length (and order!) as \code{colData(object)} (or
\code{pData2.0(object)}).
\code{which_reddim} and \code{which_pcs} will be ignored if \code{X} and 
\code{Y} are specified.}

\item{color_by}{entry of either \code{colData(object)}, \code{rowData(object)},
\code{rownames(object)}, which will be used to assign either discrete or
continous color schemes. Alternatively, you can supply a \code{list} with a
label stored in \code{list$title} and values stored in \code{list$result}
(\code{result} should be as long as \code{dim(object)[2]}. Default: \code{NULL}}

\item{shape_by}{default: \code{NULL}}

\item{size_by}{default: \code{NULL}}

\item{circle_by}{Used for encircling points of interest; default: \code{NULL}}
}
\value{
\code{ggplot2} object (a plot)

DimRedPlot object: a simple S3-based object, i.e. a list with:
\enumerate{
\item \code{plot_data}: data.frame fit for ggplot2-based plotting
\item \code{label_data}: additional stuff that may be used for labelling the
x and y axes of the final plot
\item \code{factors}: names of the plotting variables, e.g. which values
were used for defining the coloring etc.
}
}
\description{
This will generate a ggplot starting from a \code{DimRedPlot} object.
See \code{\link{get_reducedDimPlot.sce}} to start from a \code{SingleCellExperiment}
object.

Wrapper function for extracting the X and Y coordinates for the
plot as well as the variables for coloring, shape assignment etc.
This function expects a SingleCellExperiment object.
}
\details{
\emph{which_reddim} The reducedDims slot of the SingleCellExperiment object
is a list (\code{\link{SimpleList}}. To know which coordinates to extract,
the accessor for the specific reducedDims list member must be given.
}
\examples{
data("tiny_sce")
pdrp <- get_reducedDimPlot.sce(tiny_sce, which_reddim = "PCA",
                              which_pcs = c(2:3),
                              color_by = names(colData(tiny_sce))[1],
                              dim_red_type = "PCA",
                              add_cell_info = names(colData(tiny_sce)),
                              exprs_values = "log10medScaled")
# color based on the gene used during the first step
plt.DimRedPlot(pdrp)

# no color
plt.DimRedPlot(pdrp, ignore_drp_labels = "color_by")
  
# color based on condition, avoid "log10medScaled" being printed in the legend
plt.DimRedPlot(pdrp, color_by = "condition", ignore_drp_labels = "exprs_val_type")

# color based on barcode, encircle points that belong to the same condition 
plt.DimRedPlot(pdrp, color_by = "barcode", ignore_drp_labels = "exprs_val_type", circle_by = "condition")


data("tiny_sce")

# use reducedDimensions stored within the tiny_sce
drp <- get_reducedDimPlot.sce(tiny_sce, which_reddim = "PCA",
                                          which_pcs = c(2:3), 
                                          color_by = names(colData(tiny_sce))[1],
                                          dim_red_type = "PCA" )
head(drp$plot_data)
head(drp$label_data)

# specify X and Y manually
x <- reducedDim(tiny_sce, "PCA")[,"PC1", drop = FALSE]
y <- reducedDim(tiny_sce, "PCA")[,"PC2", drop = FALSE]

drp <- get_reducedDimPlot.sce(tiny_sce, X = x, Y = y)

## note the difference in the entries for drp$label_data depending on whether
## drop = FALSE is used to define x and y (if the colnames are dropped, they
## won't be part of the DimRedPlot object)
}
\seealso{
\code{\link{get_reducedDimPlot.sce}}, which generates the DimRedPlot
object that will be used here.

\code{\link{SingleCellExperiment}}, \code{\link{reducedDims}},
\code{\link{plt.DimRedPlot}}
}
