\name{github_auth}
\alias{github_auth}
\title{Authenticate with github}
\usage{
  github_auth(client_id = NULL, client_secret = NULL)
}
\arguments{
  \item{client_id}{Consumer key. can be supplied here or
  read from Options()}

  \item{client_secret}{Consumer secret. can be supplied
  here or read from Options()}
}
\description{
  Authenticate with github
}
\examples{
\dontrun{
github_auth()
}
}

