\name{github_timeseries}
\alias{github_timeseries}
\title{Create time series bar plot.}
\usage{
  github_timeseries(data = NULL)
}
\arguments{
  \item{data}{Data set to plot with.}
}
\description{
  Create time series bar plot.
}
\examples{
\dontrun{
# Run with example data set (commits from the ropensci organization account)
github_timeseries()
}
}

