% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexOfUnion.R
\name{indexOfUnion}
\alias{indexOfUnion}
\title{Index of Union}
\usage{
indexOfUnion(rocobj)
}
\arguments{
\item{roc.data}{An object returned from the roc function in the package
pROC}
}
\value{
A list which contains the name of the method and the return IOU
}
\description{
Index of Union
}
