% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_Utils.R
\name{makeSandbox}
\alias{makeSandbox}
\title{Generate Connection to original R Session}
\usage{
makeSandbox(node, ID, host = "local")
}
\arguments{
\item{node}{host to run R session on}

\item{ID}{the port ID to open socket on}

\item{host}{location of host server, local or external}
}
\description{
Generate Connection to original R Session
}
