% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_SAR_OSL.R
\name{measure_SAR_OSL}
\alias{measure_SAR_OSL}
\title{Measure an aliquot with the CW SAR OSL protocol}
\usage{
measure_SAR_OSL(aliquot, sequence, dose_rate = 0.1)
}
\arguments{
\item{aliquot}{\link{data.frame}, a set of grains that are assigned to an
aliquot (sample subset used for measurement), i.e., the result of
\link{prepare_Aliquot}.}

\item{sequence}{\link{list}, Definition of the SAR protocol.}

\item{dose_rate}{\link{numeric} value, Dose rate of the luminescence
reader, in Gy.}
}
\value{
\link[Luminescence:RLum.Analysis-class]{Luminescence::RLum.Analysis} object. Equivalent of the import result for
a real world measurement file. This object can be evaluated by functions
of the package \link{Luminescence-package}.
}
\description{
The function models the time-dependent photon counts of an aliquot
according to the specified CW SAR OSL (continuous wave, single aliquot
regenerative dose protocol for optically stimulated luminescence) sequence
and parameters. The modelling is done for each component and photon count
curves are summed to return an \link[Luminescence:RLum.Analysis-class]{Luminescence::RLum.Analysis} object as equivalent of
importing a real measurement data set to the R-package \link{Luminescence-package}.

The function uses the package \link{RLumModel-package} to perform the simulation of the
photon count curves.
}
\examples{

\dontrun{

## create dummy data set

}

}
\author{
Michael Dietze, GFZ Potsdam (Germany)
}
