% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestROCthreshold.R
\name{minPVal}
\alias{minPVal}
\title{Minimum P-value}
\usage{
minPVal(time, status, pred)
}
\arguments{
\item{time}{Follow-up time}

\item{status}{Censoring status}

\item{pred}{Predictions}
}
\value{
A list with two elements. The first element is a character string
    letting you know which result is returned. The second element is the
    threshold which minimizes the p-value
}
\description{
Provides the receiver operating characteristic (ROC) threshold
    that minimizes the P-value.
}
\examples{

time <- rexp(100, 10)
status <- rbinom(100, 1, 0.5)
pred <- runif(100, 0, 1)

minPVal(time, status, pred)

}
