% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{receiveSand}
\alias{receiveSand}
\title{Get code to evaluate from original R Session}
\usage{
receiveSand(sandboxCon)
}
\arguments{
\item{sandboxCon}{socket connection to original R session}
}
\description{
Get code to evaluate from original R Session
}
