% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox-package.R
\docType{package}
\name{sandbox-package}
\alias{sandbox-package}
\alias{sandbox}
\title{Probablistic Numerical Modelling Of Sediment Properties}
\description{
Flexible framework for definition and application of time/depth-based 
rules for sets of parameters for single grains that can be used to 
create synthetic samples, used for synthetic preparation and synthetic 
measurements.
}
\details{
\tabular{ll}{ Package: \tab sandbox\cr Type: \tab Package\cr Version: \tab
0.0.2\cr Date: \tab 2016-10-24\cr License: \tab GPL-3\cr }
}
\author{
Michael Dietze
}
\keyword{package}
