% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\docType{package}
\name{sandbox-package}
\alias{sandbox}
\alias{sandbox-package}
\title{Gamify your R experience}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Gamify learning programming in R. Get challenges and feedback on progress!
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/D-Se/sandbox}
  \item Report bugs at \url{https://github.com/D-Se/ergo/issues}
}

}
\author{
\strong{Maintainer}: Donald Seinen \email{2869830202@qq.com} (\href{https://orcid.org/0000-0002-4343-5851}{ORCID})

}
\keyword{Problem}
\keyword{are}
\keyword{born}
\keyword{internal}
\keyword{problem}
\keyword{skills}
\keyword{solving}
\keyword{solving.}
\keyword{through}
