% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox-package.R
\docType{package}
\name{sandbox-package}
\alias{sandbox}
\alias{sandbox-package}
\title{Sandbox}
\description{
Package of Miscellaneous Functions.
}
\details{
This package contains a mixed-bag of functions. Function topics include:
\itemize{

\item Calibration metrics for competing risk regression analysis
\item Other common risk prediction metrics (ROC, AUC, predictions, optimal ROC thresholds, NRI, Nagelkerke's R-sq, Hosmer-Lemeshow test)
\item Genomic tools (HWE tests, FDR)
}
}
\section{GitHub}{

\href{https://github.com/warkentinmatt/sandbox}{warkentinmatt/sandbox}
}

\author{
\itemize{

\item \strong{Matthew T. Warkentin} \email{warkentin@lunenfeld.ca}

}
}
