% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\name{sandbox}
\alias{sandbox}
\title{Run expression in a sandboxed environment and return outputs}
\usage{
sandbox(x, env = parent.frame())
}
\description{
Run expression in a sandboxed environment and return outputs
}
\examples{
\dontrun{
sandbox({
  library(tidyverse)
  testval<-22
  print(testval)
  print(testVar)
})
}
}
