% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\name{sandbox}
\alias{sandbox}
\title{Sandbox code into another R session}
\usage{
sandbox(x, sbConnection = sandboxConnectionTemplate(), env = parent.frame())
}
\arguments{
\item{x}{an expression to be exectuted, or file name of script, in the sandbox environment}

\item{sbConnection}{output of the [sandboxConnectionTemplate()]}

\item{env}{envionment that the leaked objects are to be assigned to. Assumes [global.env()]}
}
\description{
Run expression in a sandboxed environment and return outputs
}
\examples{
sandbox({
  testval<-22
  print(testval)
})
}
