% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yang_git_functions.R
\name{sum_square_fibo}
\alias{sum_square_fibo}
\title{sum up the first nth fibonacci numbers with power 2.}
\usage{
sum_square_fibo(n)
}
\arguments{
\item{n}{positive integer}
}
\value{
print a prompt with "Sum of the first nth of Fibonacci numbers of power 2:",
along with the sum square value.
}
\description{
sum up the first nth fibonacci numbers with power 2.
}
\examples{
sum_square_fibo(3)
}
