% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestROCthreshold.R
\name{topleft}
\alias{topleft}
\title{Closest to Top Left}
\usage{
topleft(sens, spec, thr)
}
\arguments{
\item{sens}{Sensitivity}

\item{spec}{Specificity}

\item{thr}{Threshold}
}
\value{
A list with two elements. The first element is a character string
    letting you know which type of result is returned. The second element is the
    threshold which corresponds to the minimized distance
}
\description{
Provides the receiver operating characteristic (ROC) threshold
    (x-axis) that minimizes the distance from the ROC curve to the  to the
    top left corner of the plot (1,1)
}
\examples{

sens <- runif(100, 0, 1)
spec <- runif(100, 0, 1)
thr <- runif(100, 0, 1)

topleft(sens, spec, thr)

}
